Correções de tipo de dados
alteração de duble para bigdecimal e alterações de valor fixos para gets

+++ src/main/java/DAL/CursoSqlDAO.java
@@ -7,6 +7,7 @@ import model.Departamento;
 import model.Resultado;
 import model.UnidadeCurricular;
 
+import java.math.BigDecimal;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.Statement;
@@ -31,7 +32,7 @@ public class CursoSqlDAO implements ICursoDAO {
                     rs.getString("depSigla")
             );
             Curso curso = new Curso(rs.getString("nome"), rs.getInt("duracao"), dep);
-            curso.setPrecoAnual(rs.getDouble("precoAnual"));
+            curso.setPrecoAnual(BigDecimal.valueOf(rs.getDouble("precoAnual")));
 
             int cursoId = rs.getInt("id");
 
@@ -116,7 +117,7 @@ public class CursoSqlDAO implements ICursoDAO {
                 ps.setString(1, curso.getNome());
                 ps.setInt(2, curso.getDuracao());
                 ps.setInt(3, depId);
-                ps.setDouble(4, curso.getPrecoAnual());
+                ps.setBigDecimal(4, curso.getPrecoAnual());
                 ps.executeUpdate();
                 try (ResultSet keys = ps.getGeneratedKeys()) {
                     if (keys.next()) idGerado[0] = keys.getInt(1);
@@ -182,7 +183,7 @@ public class CursoSqlDAO implements ICursoDAO {
                 ps.setString(1, cursoNovo.getNome());
                 ps.setInt(2, cursoNovo.getDuracao());
                 ps.setInt(3, depId);
-                ps.setDouble(4, cursoNovo.getPrecoAnual());
+                ps.setBigDecimal(4, cursoNovo.getPrecoAnual());
                 ps.setInt(5, cursoId);
                 int rows = ps.executeUpdate();
                 if (rows == 0) throw new java.sql.SQLException("Curso não encontrado para atualizar.");
